requirements.txt
python-dotenv
tweepy
Marine-Traffic-API
pytz
python-dateutil